Update patch set 1
Patch Set 1:

no worries, it's not too much of a big deal. I was able to keep f2fs stable in a unofficial cm based rom, but with cm official it worked but unfortunately not to the same standard. Will keep your comments in mind though because i am working on f2fs for devices on cm-13.0 right now

Patch-set: 1